+++ peril-settings.json
@@ -4,14 +4,15 @@
     "disable_github_check": true
   },
   "rules": {
-    "pull_request, pull_request.labeled, pull_request.unlabeled": "org/common.ts"
+    "pull_request, pull_request.labeled, pull_request.unlabeled": "markpar/peril-settings@org/common.ts"
   },
   "repos" : {
     "markpar/WordPress-iOS" : {
       "pull_request, pull_request.labeled, pull_request.unlabeled": "markpar/peril-settings@org/ios.ts"
     },
     "markpar/WordPress-Android" : {
-      "pull_request, pull_request.labeled, pull_request.unlabeled": "markpar/peril-settings@org/android.ts"
+      "pull_request, pull_request.labeled, pull_request.unlabeled": "markpar/peril-settings@org/android.ts",
+      "pull_request.opened, pull_request.edited": "markpar/peril-settings@org/prForSubtreeChanges.ts"
     },
     "markpar/woocommerce-ios" : {
       "pull_request, pull_request.labeled, pull_request.unlabeled": "markpar/peril-settings@org/ios.ts"
